Refactor all products views to functions and sort them by topic into files
