+++ hagrid/products/templates/dashboard.html
@@ -1,8 +1,8 @@
 {% extends "base.html" %}
 
 {% block content %}
-    <h1 class="title">Welcome to FOC</h1>
-    <h1 class="center">The 38c3 fashion operating center</h1>
+    <h1 class="title">Welcome to C3FOC</h1>
+    <h1 class="center">The 38C3 fashion operation center</h1>
 
     {% if dashboard_text %}
         <div class="dashboard-row">
