+++ hagrid/products/views/__init__.py
@@ -1,0 +1,12 @@
+from .config import (
+    products_config_overview_view,
+    variation_availability_config_view,
+    variation_availability_event_list_view,
+    variation_config_view,
+)
+from .dashboard import dashboard_table_view, dashboard_view
+from .variation_count import (
+    variation_count_overview_view,
+    variation_count_success_view,
+    variation_count_view,
+)
