+++ hagrid/urls.py
@@ -17,12 +17,11 @@ from django.contrib import admin
 from django.urls import include, path
 from django.conf.urls.static import static
 from django.conf import settings
-from django.views.decorators.cache import cache_page
-from hagrid.products.views import DashboardView, DashboardTableView
+from hagrid.products.views import dashboard_view, dashboard_table_view
 
 urlpatterns = [
-    path('', cache_page(10)(DashboardView.as_view()), name='dashboard'),
-    path('table', cache_page(10)(DashboardTableView.as_view()), name='dashboard_table'),
+    path('', dashboard_view, name='dashboard'),
+    path('table', dashboard_table_view, name='dashboard_table'),
     path('admin/', admin.site.urls),
     path('products/', include('hagrid.products.urls')),
     path('reservations/', include('hagrid.reservations.urls')),
